% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access_token.R
\name{get_access_token}
\alias{get_access_token}
\title{Get token to access EpiCollect entries}
\usage{
get_access_token(client_id, client_secret)
}
\arguments{
\item{client_secret}{}
}
\description{
Get token to access EpiCollect entries
}
