% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_epicollect.R
\name{read_epicollect}
\alias{read_epicollect}
\title{Get EpiCollect entries in tidy format}
\usage{
read_epicollect(project_slug, token)
}
\arguments{
\item{token}{}
}
\description{
Get EpiCollect entries in tidy format
}
